refactor(DEVING-25): 회원가입 로직과 ui 분리 리펙토링
