+++ src/app/signup/page.tsx
@@ -1,18 +1,4 @@
-'use client';
-
-import { Button } from '@/components/ui/Button';
-import { Input } from '@/components/ui/Input';
-import {
-  useEmailCheckMutation,
-  useNameCheckMutation,
-  useSignupMutation,
-} from '@/hooks/mutations/useUserMutation';
-import Link from 'next/link';
-import { useRouter } from 'next/navigation';
-import { useEffect, useState } from 'react';
-import { useForm } from 'react-hook-form';
-
-import { ChipContainer } from './components/ChipContainer';
+import SignupForm from './components/SignupForm';
 
 export interface ISignupFormData {
   name: string;
@@ -23,285 +9,9 @@ export interface ISignupFormData {
 }
 
 export default function Signup() {
-  const {
-    register,
-    handleSubmit,
-    watch,
-    trigger,
-    setError,
-    setValue,
-    formState: { errors, dirtyFields },
-  } = useForm<ISignupFormData>({
-    mode: 'onBlur',
-    defaultValues: {
-      position: '',
-    },
-  });
-
-  /**
-   * TODO
-   * 포커스 1초 되 유효성 검사 - 닉네임, 이메일, 비밀번호, 비밀번호 확인
-   */
-
-  /**
-   * TODO
-   * 중복확인 플로우 체크
-   * 1. 처음에는 중복확인 버튼 비활성화(''인 경우)
-   * 2. 입력이 있다면 중복확인 버튼 활성화
-   *  - 입력 중에는 비활성화
-   * 3. 중복확인 성공시
-   *  - 해당 인풋 성공 표시(보더 색 변경), disabled 없이 그대로 유지
-   *  - 중복확인 버튼 비활성화
-   * 3.1. 중복확인 성공 후 다시 입력 시
-   *  - 중복확인 버튼 활성화
-   *  - 해당 인풋 성공 처리 취소
-   * 4. 중복확인 실패시
-   *  - 해당 인풋 에러메시지 표시
-   *  - 중복확인 버튼은 계속 활성화
-   */
-
-  // 닉네임 중복 체크 확인.
-  // 1. 처음에는 중복확인 버튼 비활성화
-  const [isNameCheck, setIsNameCheck] = useState(false);
-  const [isEmailCheck, setIsEmailCheck] = useState(false);
-
-  // 회원가입 버튼 활성화 여부
-  const router = useRouter();
-
-  // 2. 입력이 있다면 중복확인 버튼 활성화
-  useEffect(() => {
-    setIsNameCheck(false);
-  }, [watch('name')]);
-
-  useEffect(() => {
-    setIsEmailCheck(false);
-  }, [watch('email')]);
-
-  // 3. 중복확인 로직 수행
-  const { mutate: nameCheckMutate } = useNameCheckMutation({
-    onSuccessCallback: () => setIsNameCheck(true),
-    onErrorCallback: () =>
-      setError('name', {
-        type: 'checkFail',
-        message: '이미 존재하는 닉네임입니다.',
-      }),
-  });
-
-  const { mutate: emailCheckMutate } = useEmailCheckMutation({
-    onSuccessCallback: () => setIsEmailCheck(true),
-    onErrorCallback: () =>
-      setError('email', {
-        type: 'checkFail',
-        message: '이미 존재하는 이메일입니다.',
-      }),
-  });
-
-  const handleNameCheck = () => {
-    const name = watch('name');
-    nameCheckMutate(name);
-    trigger('name');
-  };
-
-  const handleEmailCheck = () => {
-    const email = watch('email');
-    emailCheckMutate(email);
-    trigger('email');
-  };
-
-  const { mutate: singupMutate } = useSignupMutation({
-    onSuccessCallback: () => router.push('/login'),
-  });
-
-  const onSubmit = (data: ISignupFormData) => {
-    console.log('회원가입 데이터: ', data);
-
-    /**
-     * TODO
-     * - 닉네임 중복검사 했는지 확인
-     * - 이메일 중복검사 했는지 확인
-     * - position 선택했는지 확인
-     * - 비밀번호 입력했는지 확인
-     * - 비밀번호 확인 입력했는지 확인
-     */
-
-    if (!isNameCheck) {
-      setError('name', {
-        type: 'nameCheck',
-        message: '닉네임 중복확인이 필요합니다.',
-      });
-    }
-    if (!isEmailCheck) {
-      setError('email', {
-        type: 'emailCheck',
-        message: '이메일 중복확인이 필요합니다.',
-      });
-    }
-    if (watch('position') === '') {
-      setError('position', {
-        type: 'positionCheck',
-        message: '포지션을 선택해 주세요.',
-      });
-    }
-
-    if (Object.keys(errors).length) {
-      return;
-    }
-    singupMutate(data);
-  };
-
-  // 포지션 클릭
-  const handleClickPosition = (value: string) => {
-    setValue('position', value);
-    trigger('position');
-  };
-
   return (
     <div className="flex min-h-screen items-center justify-center">
-      <form
-        onSubmit={handleSubmit(onSubmit)}
-        className="flex w-[544px] flex-col rounded-[16px] bg-BG_2 p-[40px]"
-      >
-        <div>
-          <h2 className="typo-head2 mb-[40px] text-center text-white">
-            회원가입
-          </h2>
-          <div className="flex flex-col gap-[24px]">
-            <div className="flex flex-col gap-[8px]">
-              <label htmlFor="name" className="typo-head3 text-Cgray700">
-                닉네임
-              </label>
-              <div className="flex flex-row  gap-[8px]">
-                <Input
-                  id="name"
-                  className="h-full"
-                  placeholder="닉네임을 입력해주세요."
-                  {...register('name', { required: '닉네임을 입력해주세요.' })}
-                  errorMessage={errors.name?.message}
-                  state={isNameCheck ? 'success' : 'default'}
-                />
-                <Button
-                  disabled={isNameCheck}
-                  variant={'outline'}
-                  size={'sm'}
-                  className="h-[50px]"
-                  type="button"
-                  onClick={handleNameCheck}
-                >
-                  중복확인
-                </Button>
-              </div>
-            </div>
-
-            <div className="flex flex-col gap-[8px]">
-              <label htmlFor="id" className="typo-head3 text-Cgray700">
-                이메일
-              </label>
-              <div className="flex flex-row  gap-[8px]">
-                <Input
-                  id="email"
-                  className=" h-full"
-                  placeholder="이메일을 입력해주세요."
-                  {...register('email', {
-                    required: '이메일을 입력해주세요.',
-                    pattern: {
-                      value: /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/,
-                      message: '올바른 이메일 형식이 아닙니다.',
-                    },
-                  })}
-                  state={isEmailCheck ? 'success' : 'default'}
-                  errorMessage={errors.email?.message}
-                />
-                <Button
-                  disabled={isEmailCheck}
-                  variant={'outline'}
-                  size={'sm'}
-                  className="h-[50px]"
-                  onClick={handleEmailCheck}
-                  type="button"
-                >
-                  중복확인
-                </Button>
-              </div>
-            </div>
-
-            <div className="flex flex-col gap-[8px]">
-              <label htmlFor="id" className="typo-head3 text-Cgray700">
-                포지션
-              </label>
-              <ChipContainer
-                position={watch('position')}
-                setPosition={handleClickPosition}
-              />
-              {errors.position?.message && (
-                <p className="typo-caption1 mt-[10px] px-[10px] text-warning">
-                  포지션을 선택해 주세요.
-                </p>
-              )}
-              <input
-                type="hidden"
-                {...register('position', {
-                  required: '포지션을 선택해 주세요.',
-                })}
-              />
-            </div>
-
-            <div className="flex flex-col gap-[8px]">
-              <label htmlFor="password" className="typo-head3 text-Cgray700">
-                비밀번호
-              </label>
-              <Input
-                id="password"
-                type="password"
-                placeholder="비밀번호를 입력해주세요."
-                {...register('password', {
-                  required: '비밀번호를 입력해주세요.',
-                  minLength: {
-                    value: 6,
-                    message: '비밀번호는 최소 6자 이상이어야 합니다.',
-                  },
-                  pattern: {
-                    value: /^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\d]{6,}$/,
-                    message:
-                      '비밀번호는 영어와 숫자 포함 6자 이상이어야 합니다.',
-                  },
-                })}
-                state={dirtyFields.password ? 'success' : 'default'}
-                errorMessage={errors.password?.message}
-              />
-            </div>
-
-            <div className="flex flex-col gap-[8px]">
-              <label htmlFor="pw" className="typo-head3 text-Cgray700">
-                비밀번호 확인
-              </label>
-              <Input
-                id="passwordCheck"
-                type="password"
-                placeholder="비밀번호를 입력해주세요."
-                {...register('passwordCheck', {
-                  required: '비밀번호를 입력해주세요.',
-                  validate: (value) =>
-                    value === watch('password') ||
-                    '비밀번호가 일치하지 않습니다.',
-                })}
-                state={dirtyFields.passwordCheck ? 'success' : 'default'}
-                errorMessage={errors.passwordCheck?.message}
-              />
-            </div>
-          </div>
-        </div>
-        <div className="mb-[20px] mt-[48px] flex flex-col">
-          <Button type="submit" className="w-full">
-            회원가입
-          </Button>
-        </div>
-        <div className="flex justify-between">
-          <p className="text-Cgray700">이미 회원이신가요?</p>
-          <Link href="/login" className="text-main underline">
-            로그인
-          </Link>
-        </div>
-      </form>
+      <SignupForm />
     </div>
   );
 }
